Refactor button element
