+++ src/components/_general/button/button.scss
@@ -7,49 +7,48 @@
     overflow: hidden;
     height: $button__size;
     width: $button__size;
-    &__selected {
+    &--selected {
         opacity: 0.3;
     }
-    &__minus {
+    &--minus {
 
     }
-    &__reset-word {
+    &--reset-word {
         margin: 0 1px;
         width: 75px;
     }
-    &__plus {
+    &--plus {
 
     }
-    &__count {
+    &--count {
         height: $count-button__size;
         width: $count-button__size;
         margin: $margin 1px;
     }
-    &__reset-all-words {
+    &--reset-all-words {
         height: $count-button__size;
         width: 100px;
         margin: $margin 1px;
     }
-    &__pass-text {
-        height: $pass-text-button__height;
-        width: $pass-text-button__width;
-        min-width: $pass-text-button__width;
-    }
-    &__pass-box {
+    &--caps {
         height: $pass-box-button__height;
         width: $pass-box-button__width;
         min-width: $pass-box-button__width;
         margin: 1px 0;
-        &--caps {
-            border-top-right-radius: 0;
-            border-bottom-right-radius: 0;
-        }
-        &--copy {
-            border-top-left-radius: 0;
-            border-bottom-left-radius: 0;
-            &-icon {
-                @include fa($fa-clipboard, $bg-color);
-            }
+
+        border-top-right-radius: 0;
+        border-bottom-right-radius: 0;
+    }
+    &--copy {
+        height: $pass-box-button__height;
+        width: $pass-box-button__width;
+        min-width: $pass-box-button__width;
+        margin: 1px 0;
+
+        border-top-left-radius: 0;
+        border-bottom-left-radius: 0;
+        i {
+            @include fa($fa-clipboard, $bg-color, $pass-box-button__height - 10px);
         }
     }
 }
