+++ src/components/template.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import './__.scss';
 
 const Component = (props) => (
     <div>
