+++ src/components/wordcount_buttons/wordcount_buttons.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import {observer, inject} from 'mobx-react';
-import cn from 'classnames';
 
-import Button from '../general/button/button';
+import Button from '../_general/button/button';
 
 import './wordcount_buttons.scss';
 
@@ -10,16 +9,15 @@ const WordCountButtonsContainer = inject('wordStore')(observer(({wordStore}) => 
     const buttons = [];
     for (let i = 1; i <= wordStore.maxWords; i++) {
         buttons[i] =
-            <Button className={cn("btn btn__count", {"btn__selected": i === wordStore.usedWords})}
-                    key={"wcb_" + i} onClick={wordStore.handleUsedWords[i - 1]} text={i}/>;
+            <Button type="count" selected={i === wordStore.usedWords} text={i}
+                    key={"wcb_" + i} onClick={wordStore.handleUsedWords[i - 1]}/>;
     }
     return (
         <div className="buttons-container">
             {buttons}
             <div>
-                <Button className="btn btn__reset-all-words"
-                        onClick={wordStore.handleResetAllWords}
-                        text="Изменить все"/>
+                <Button type="reset-all-words" text="Изменить все"
+                        onClick={wordStore.handleResetAllWords}/>
             </div>
         </div>
     );
