+++ src/vars.scss
@@ -26,6 +26,8 @@ $app-menu__height: 30px;
 $app-menu__color: $bg-color;
 $app-menu__bg-color: $text-color;
 
+$pass-font-size: 30px;
+
 // BUTTONS
 $button__radius: 2px;
 $button__size: 20px;
